pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.0"
dependencies = [
    "ruff>=0.6.2",
    "fastapi>=0.112.1",
    "uvicorn>=0.30.6",
    "sqlalchemy>=2.0.32",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.9",
    "greenlet>=3.0.3",
    "python-jose>=3.3.0",
    "httpx>=0.27.0",
    "protobuf>=4.25.4",
    "trio>=0.26.2",
    "logfire[fastapi]==3.5.3",
    "opentelemetry-instrumentation-sqlalchemy>=0.47b0",
    "opentelemetry-instrumentation-asyncio>=0.47b0",
    "clickhouse-connect>=0.7.19",
    "boto3>=1.35.5",
    "scalar-fastapi>=1.0.3",
    "cryptography>=43.0.1",
    "google-cloud-pubsub>=2.23.1",
    "websockets>=13.0.1",
    "python-multipart>=0.0.9",
    "modal==1.0.1",
    "sqlmodel>=0.0.22",
    "aioboto3>=7.0.0",
    "upstash-redis>=1.2.0",
    "huggingface-hub>=0.26.0",
    "tabulate>=0.9.0",
    "comfy-models==0.1.5",
    "fal-client>=0.5.6",
    "stripe>=11.3.0",
    "resend>=2.5.1",
    "discord-py>=2.4.0",
    "dub==0.18.2",
    "logtail-python>=0.3.3",
    "nanoid>=2.0.0",
    "google-auth>=2.40.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
]
